DSAI3202-Phase1/requirements.txt
pandas==2.1.0
numpy==1.24.3
librosa==0.10.1
matplotlib==3.7.2
seaborn==0.13.0
scikit-learn==1.3.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
pyarrow==13.0.0
soundfile==0.12.1
tqdm==4.66.1
